Added support for specifying cluster type in the cloud config file (-c <file> on cmdline): strings Nimbus, OpenNebula, and Eucalyptus currently supported.
The config reader parses the config line accurately, and creates a Cluster object of the subclass type correpsonding to the config's type string. That is, if the type is Nimbus in the config file, a NimbusCluster object will be created to hold that line's information.
Currently, only NimbusCluster exists in cloud_management, so it is the only cloud_type fully supported. OpenNebula and Eucalyptus types in the config file result in the creation of standard superclass Cluster objects.