Fixed a final bug in blankspace functionality and updated checkout/return methods to consider cluster storage.
Bug fixed in blankspace: the mount scheme is now sda for the root partition and sdb for the blankspace partition. Note that the blankspace parition must be mounted manually inside the image (this could be set up to be done automatically, with a line in fstab of the VMs).
Checkout and return now checkout and return the storage required by a VM. Notably, the storage field in the resource config file for clusters should be the sum total of all storage available on the worker nodes of the cluster.