Fix to stop a stale classad refresh causing a VM to be shutdown.
Shutdown caused by condor dropping a stale ad of a VM and CS seeing the missing ad of a valid Running VM after the register time limit triggering a shutdown.
Resetting the last_state_change and condorname will give the VM the full duration of the time_limit to re-register with condor before it will be shutdown.