Adjust resources on Nimbus clouds if vm creation errors related to lack of slots or memory
Issues with CS cycling on creation a VM on a cloud that it is out of resources but CS has more resource configured.
Parses out the error string from Nimbus return to determine type of creation error and if it's network slot or memory related will adjust the resources within CS to reflect the state of cloud rather than the configured value.